# proto/openapi.yaml
openapi: 3.0.3
info:
  title: Banking Platform API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
  /api/auth/register:
    post:
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Registered user and wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/auth/login:
    post:
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Access and refresh tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/auth/refresh:
    post:
      operationId: refresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Rotated tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/auth/pin/verify:
    post:
      operationId: verifyPin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PinVerifyRequest"
      responses:
        "200":
          description: PIN verified for the authenticated user
  /api/auth/password/otp:
    post:
      operationId: requestPasswordResetOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier]
              properties:
                identifier:
                  type: string
                  description: Registered email or phone number. The OTP is sent to the account email.
      responses:
        "200":
          description: OTP request accepted. Response does not disclose whether the identifier exists.
  /api/auth/password/reset:
    post:
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier, otp, newPassword]
              properties:
                identifier:
                  type: string
                  description: Registered email or phone number
                otp:
                  type: string
                  pattern: "^\\d{6}$"
                  description: Password reset OTP sent to the account email
                newPassword:
                  type: string
                  minLength: 6
      responses:
        "200":
          description: Password reset; user can log in with the new password
        "400":
          description: Missing/invalid new password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Invalid, expired, locked, or missing OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/accounts/lookup:
    get:
      operationId: lookupAccount
      parameters:
        - name: email
          in: query
          schema:
            type: string
        - name: phone
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Account lookup result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountDetails"
  /api/accounts/{id}:
    get:
      operationId: getAccountDetails
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Account details including registered owner contact information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountDetails"
  /api/accounts/{id}/balance:
    get:
      operationId: getBalance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Decimal string balance
  /api/accounts/{id}/deposit:
    post:
      operationId: deposit
      description: Admin-only mock deposit against the SYSTEM CASH_CLEARING account.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          schema:
            type: string
      responses:
        "200":
          description: Deposit ledger result
  /api/accounts/{id}/withdraw:
    post:
      operationId: withdraw
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          schema:
            type: string
      responses:
        "200":
          description: Withdraw ledger result
...
  /api/admin/accounts:
    get:
      operationId: adminListAccounts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated enriched admin account list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccountPage"
  /api/admin/accounts/{id}:
    get:
      operationId: adminGetAccount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Enriched admin account detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccount"
  /api/admin/transactions:
    get:
      operationId: adminListTransactions
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
        - name: status
          in: query
          schema:
            type: string
        - name: accountId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated admin transaction list
  /api/admin/accounts/{id}/suspend:
    post:
      operationId: adminSuspendAccount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Suspended account
...
  /api/admin/audit:
    get:
      operationId: adminListAudit
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 15
      responses:
        "200":
          description: Paginated admin audit list
...
components:
  schemas:
    ErrorResponse:
      type: object
      required: [code, message, traceId]
      properties:
        code:
          type: string
        message:
          type: string
        traceId:
          type: string
    RegisterRequest:
      type: object
      required: [email, password, pin]
      properties:
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true
        password:
          type: string
          minLength: 6
          writeOnly: true
        pin:
          type: string
          pattern: "^\\d{6}$"
          writeOnly: true
        currency:
          type: string
          default: VND
          enum: [VND]
    LoginRequest:
      type: object
      required: [identifier, password]
      properties:
        identifier:
          type: string
          description: Registered email or phone number
        password:
          type: string
          writeOnly: true
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          writeOnly: true
    PinVerifyRequest:
      type: object
      required: [pin]
      properties:
        pin:
          type: string
          pattern: "^\\d{6}$"
          writeOnly: true
    AuthResponse:
      type: object
      required: [accessToken, refreshToken, userId, roles]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        userId:
          type: string
          format: uuid
        accountId:
          type: string
          format: uuid
          nullable: true
        roles:
          type: array
          items:
            type: string
    MoneyString:
      type: string
      pattern: "^-?[0-9]+(\\.[0-9]+)?$"
...
    AdminAccount:
      type: object
      required: [id, currency, kind, status, balance, createdAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
          nullable: true
        email:
          type: string
          nullable: true
        phoneNumber:
          type: string
          nullable: true
        code:
          type: string
          nullable: true
        currency:
          type: string
        kind:
          type: string
          enum: [USER, SYSTEM]
        status:
          type: string
        balance:
          $ref: "#/components/schemas/MoneyString"
        createdAt:
          type: string
          format: date-time
    AccountDetails:
      type: object
      required: [id, code, currency, kind, status, version, createdAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        code:
          type: string
        currency:
          type: string
        kind:
          type: string
          enum: [USER, SYSTEM]
        status:
          type: string
        version:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
    AdminAccountPage:
      type: object
      required: [items, page, size, totalElements, totalPages]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminAccount"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
...